Refactor ProfileList and move FileUploadForm
