+++ src/js/components/admin/common/FileUploadForm.js
